+++ src/forms/createModelForm.tsx
@@ -6,12 +6,17 @@ import {
   MenuItem
 } from "@blueprintjs/core";
 import "@blueprintjs/core/lib/css/blueprint.css";
-import { ItemPredicate, ItemRenderer, Suggest } from "@blueprintjs/select";
 import Form, { FormComponentProps } from "antd/lib/form";
 import axios from "axios";
 import * as React from "react";
 import { connect } from "react-redux";
 import { API_ROOT } from "../api-config";
+import {
+  filterString,
+  renderCreateItemOption,
+  renderStringItem,
+  StringSuggest
+} from "./formUtils";
 import "./login.scss";
 
 export interface IModelState {
@@ -34,82 +39,8 @@ export interface IModelObject {
   comment: string | undefined;
 }
 
-function escapeRegExpChars(text: string) {
-    // eslint-disable-next-line
-  return text.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1");
-}
 const FormItem = Form.Item;
 
-export const filterVendor: ItemPredicate<string> = (
-  query,
-  vendor,
-  _index,
-  exactMatch
-) => {
-  const normalizedTitle = vendor.toLowerCase();
-  const normalizedQuery = query.toLowerCase();
-
-  if (exactMatch) {
-    return normalizedTitle === normalizedQuery;
-  } else {
-    return normalizedTitle.indexOf(normalizedQuery) >= 0;
-  }
-};
-function highlightText(text: string, query: string) {
-  let lastIndex = 0;
-  const words = query
-    .split(/\s+/)
-    .filter(word => word.length > 0)
-    .map(escapeRegExpChars);
-  if (words.length === 0) {
-    return [text];
-  }
-
-  const regexp = new RegExp(words.join("|"), "gi");
-  const tokens: React.ReactNode[] = [];
-  while (true) {
-    const match = regexp.exec(text);
-    if (!match) {
-      break;
-    }
-    const length = match[0].length;
-    const before = text.slice(lastIndex, regexp.lastIndex - length);
-    if (before.length > 0) {
-      tokens.push(before);
-    }
-    lastIndex = regexp.lastIndex;
-    tokens.push(<strong key={lastIndex}>{match[0]}</strong>);
-  }
-  const rest = text.slice(lastIndex);
-  if (rest.length > 0) {
-    tokens.push(rest);
-  }
-  return tokens;
-}
-export const renderVendor: ItemRenderer<string> = (
-  vendor,
-  { handleClick, modifiers, query }
-) => {
-  if (!modifiers.matchesPredicate) {
-    return null;
-  }
-  return <MenuItem text={highlightText(vendor, query)} onClick={handleClick} />;
-};
-export const renderCreateFilmOption = (
-  query: string,
-  active: boolean,
-  handleClick: React.MouseEventHandler<HTMLElement>
-) => (
-  <MenuItem
-    icon="add"
-    text={`Use"${query}"`}
-    active={active}
-    onClick={handleClick}
-    shouldDismissPopover={false}
-  />
-);
-
-const VendorSuggest = Suggest.ofType<string>();
 export class CreateModelForm extends React.Component<
   CreateModelFormProps & FormComponentProps,
   IModelState & IModelObject
@@ -185,7 +116,7 @@ export class CreateModelForm extends React.Component<
         >
           <h2>Add a New Model</h2>
           <FormGroup label="Vendor">
-            <VendorSuggest
+            <StringSuggest
               inputValueRenderer={(vendor: string) => vendor}
               items={this.state.vendors}
               onItemSelect={(vendor: string) =>
@@ -193,11 +124,11 @@ export class CreateModelForm extends React.Component<
                   vendor: vendor
                 })
               }
-              createNewItemRenderer={renderCreateFilmOption}
+              createNewItemRenderer={renderCreateItemOption}
               createNewItemFromQuery={(vendor: string) => vendor}
               // onItemSelect={() => {}}
-              itemRenderer={renderVendor}
-              itemPredicate={filterVendor}
+              itemRenderer={renderStringItem}
+              itemPredicate={filterString}
               noResults={<MenuItem disabled={true} text="No results." />}
             />
           </FormGroup>
